requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
dill==0.3.8
distro==1.9.0
Farama-Notifications==0.0.4
filelock==3.13.4
Flask==2.1.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.2
greenlet==3.0.3
gym==0.26.2
gym-notices==0.0.8
gym-sokoban==0.0.6
gymnasium==0.29.1
h11==0.14.0
h5py==3.11.0
hashids==1.3.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.47
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mementos==1.3.1
minigrid==2.3.1
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
openai==1.17.1
opencv-python==4.9.0.80
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
progressbar2==4.4.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
pygame==2.5.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-git-info==0.8.3
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scipy==1.13.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
sympy==1.12
TatSu==5.12.0
tenacity==8.2.3
termcolor==2.4.0
thinc==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tornado==6.4
tqdm==4.66.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
visdom==0.2.4
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.7.0
Werkzeug==2.0.3
yarl==1.9.4
